feat: Complete TrialMatch ECS stack refactoring with 90% size reduction
🎉 Major architectural transformation completed successfully!

BREAKING CHANGE: TrialMatchEcsStack completely restructured

## What Changed
- Refactored monolithic 1,731-line stack to modular 174-line architecture
- Extracted 13 specialized services with single responsibilities
- Implemented 6 builder classes for resource construction
- Added comprehensive configuration management system
- Created export and manager utilities for cross-stack operations

## New Architecture
- **Services/**: SecretManager, EcrRepositoryManager, EcsServiceFactory, LoggingManager, OutputExporter
- **Builders/**: TaskDefinitionBuilder, ContainerDefinitionBuilder, EnvironmentVariableBuilder, HealthCheckBuilder, PortMappingBuilder, IamRoleBuilder, SecurityGroupManager
- **Configuration/**: ConfigurationLoader, ConfigurationModels
- **Exporters/**: StackOutputExporter
- **Managers/**: EnvironmentVariableManager

## Benefits Achieved
✅ 90% reduction in main stack size (1,731 → 174 lines)
✅ All 330 tests passing (259 unit + 71 integration)
✅ Improved maintainability through single responsibility principle
✅ Enhanced testability with independent service testing
✅ Maximum reusability across different stacks
✅ Clear separation of concerns and loose coupling

## Testing
- All unit tests pass: 259/259 ✅
- All integration tests pass: 71/71 ✅
- TrialMatch-specific tests pass: 33/33 ✅
- CDK synthesis validation successful ✅

## Documentation
- Added comprehensive README.md explaining new architecture
- Documented all service interfaces and usage patterns
- Created deployment guide for refactored architecture
- Added architecture diagrams and component descriptions

This refactoring transforms TrialMatch from a hard-to-maintain monolithic stack into a modern, scalable, service-oriented architecture following AWS best practices.
